fix: resolve Go idiomatics issues — wg.Go() compile error, os receiver shadowing, generate.go split
- Fix wg.Go() compile error in embed.go: replace with wg.Add(1) + go func(idx, ch)
- Fix os receiver shadowing in object.go: rename os → s, remove osStderr
- Split generate.go (2186 lines) into 5 files: stream.go, tools.go, result.go, stop.go
- Fix highestInflightStep redundant initialization in both streaming and non-streaming paths
- Add defensive type assertion in schema.go with , ok idiom
- Return error for empty tool names in buildToolMap
- Document retryable string-matching fallback as best-effort
- Replace slices.Concat with append in result.go for memory efficiency
- Add 4 regression tests to prevent future regressions
