+++ docs/review-code-fixes-postfix-20260613.md
@@ -1,0 +1,128 @@
+# Code Review: Go Idiomatics Fixes — Post-Fix Review (dev-admin-QuickStart-202606132143)
+
+**Type**: Code Review
+**Reviewer**: Senior Reviewer (Agent)
+**Date**: 2026-06-13
+**Overall Assessment**: **Approved** — all fixes correct, tests comprehensive, no regressions detected
+
+---
+
+## Changes Reviewed
+
+### Production Code
+
+| File | Change | Lines | Status |
+|------|--------|-------|--------|
+| `schema.go` | Warning 2: bare type assertion → `, ok` idiom with fall-through | 46-54 | ✅ Correct |
+| `embed.go` | Critical 1: `wg.Go()` → `wg.Add(1)` + `go func(idx, ch)` | 163-182 | ✅ Correct |
+| `object.go` | Critical 2: `os` receiver → `s`, removed `osStderr` | 168+ | ✅ Correct |
+| `generate.go` | Major 4: `highestInflightStep := 1` (streaming path) | 316 | ✅ Correct |
+| `generate.go` | Minor: `highestInflightStep := 1` (non-streaming path) | 736 | ✅ Correct (reviewer fix) |
+| `tools.go` | Warning 3: empty tool name → error | 53-54 | ✅ Correct |
+| `retry.go` | Warning 4: best-effort doc for `retryable` fallback | 46-50 | ✅ Correct |
+| `result.go` | Suggestion 1: `slices.Concat` → `append` (2 sites) | 86, 115 | ✅ Correct (reviewer fix) |
+| `tools.go` | Suggestion 1: `slices.Concat` (1 site) | 430 | ✅ Correct (unchanged) |
+
+### Test Code
+
+| File | Test | Purpose | Status |
+|------|------|---------|--------|
+| `schema_test.go` | `TestSchemaFrom_CacheBadValueDoesNotPanic` | Bad cache value doesn't panic | ✅ Correct |
+| `generate_test.go` | `TestBuildToolMap_EmptyToolName` (updated) | Empty tool name returns error | ✅ Correct |
+| `embed_test.go` | `TestEmbedMany_ParallelChunksAllCollectedInOrder` | Parallel chunks order preserved | ✅ Correct |
+| `object_test.go` | `TestObjectStream_ReceiverRenameRegression` | Receiver rename doesn't break | ✅ Correct |
+
+---
+
+## Critical Issues
+(None)
+
+## Warnings
+(None)
+
+## Suggestions
+
+### Suggestion 1: `tools.go:430` — `slices.Concat` vs `append`
+- **Location**: `tools.go:430`
+- **Severity**: Info
+- **Category**: Code Quality / Consistency
+- **Description**: `msgs = slices.Concat(msgs, toolMsgs)` still uses `slices.Concat`, while `result.go:86` and `result.go:115` were changed to `append(..., ...)`. For consistency, consider `append(msgs, toolMsgs...)`. However, `slices.Concat` is valid here since both arguments are non-trivial slices (unlike the `result.go` cases where one argument was nil or single-element).
+- **Suggestion**: Optional — change to `append(msgs, toolMsgs...)` for consistency, or leave as-is.
+
+### Suggestion 2: `object_test.go:2402` — `os.ReadFile("/dev/null")` in test
+- **Location**: `object_test.go:2402`
+- **Severity**: Info
+- **Category**: Testing
+- **Description**: The test uses `os.ReadFile("/dev/null")` to prove the `os` package is accessible. This succeeds silently on Linux (returns empty content), which is fine. On systems without `/dev/null`, it returns an error (handled by `t.Log`). This is acceptable for a regression test.
+- **Suggestion**: No action needed.
+
+---
+
+## Positive Highlights
+
+1. **schema.go type assertion fix** — The `, ok` idiom with fall-through to regenerate is the correct defensive approach. Handles both cache misses and corrupted cache values gracefully.
+
+2. **embed.go parallel path** — `wg.Add(1)` + `go func(idx int, ch []string)` correctly captures loop variables by value. Semaphore pattern with `select` for context cancellation is clean and race-free.
+
+3. **Comprehensive regression tests** — Each critical fix has a corresponding regression test:
+   - `TestSchemaFrom_CacheBadValueDoesNotPanic` — stores corrupted value, verifies no panic
+   - `TestEmbedMany_ParallelChunksAllCollectedInOrder` — 8 chunks, concurrency cap 3, verifies order
+   - `TestObjectStream_ReceiverRenameRegression` — exercises all receiver methods, proves `os` resolves to package
+
+4. **tools.go empty name error** — Returns `fmt.Errorf("goai: tool name must not be empty")` following project's error naming convention (`"goai: "` prefix).
+
+5. **retry.go documentation** — Best-effort comment accurately describes string-matching as a safety net, not primary mechanism.
+
+6. **generate.go consistency** — Both streaming (line 316) and non-streaming (line 736) paths now initialize `highestInflightStep := 1` explicitly.
+
+7. **result.go idiomatic Go** — `append(msgs, finalMsg...)` and `append(reasoning, parts...)` are more idiomatic than `slices.Concat` for these cases.
+
+---
+
+## Review Checklist
+
+### Correctness
+- [x] All critical/major fixes address the root cause
+- [x] Edge cases handled (cache corruption, empty tool names, parallel chunk ordering)
+- [x] Error paths handled (empty tool name returns error, context cancellation in embed)
+- [x] Tests verify the fixes and would catch regressions
+- [x] No side effects or behavioral changes beyond the fixes
+
+### Readability
+- [x] Variable names clear and consistent
+- [x] Comments explain non-obvious logic
+- [x] No unnecessary complexity introduced
+
+### Architecture
+- [x] Changes follow existing patterns
+- [x] Module boundaries maintained (all files `package goai`)
+- [x] No circular dependencies introduced
+- [x] File split (generate.go → 5 files) follows SRP
+
+### Security
+- [x] No secrets exposed
+- [x] No injection vulnerabilities
+- [x] Error messages don't leak sensitive data
+- [x] No new external dependencies
+
+### Performance
+- [x] No N+1 patterns introduced
+- [x] Parallel embedding path correctly bounded by semaphore
+- [x] `append(...)` in result.go reuses backing array capacity (improvement over `slices.Concat`)
+- [x] No unnecessary allocations
+
+### Testing
+- [x] Regression tests for all critical/major fixes
+- [x] Test names are descriptive and follow project conventions
+- [x] Edge cases covered (bad cache values, empty tool names, parallel ordering)
+- [x] Tests use mocks appropriately (mockModel, mockEmbeddingModel)
+
+### Imports
+- [x] No unused imports (verified `slices` still used by `slices.Clone` in result.go:102)
+- [x] Import organization follows project conventions
+
+---
+
+## Verdict
+
+**Approve** — All critical and major issues resolved correctly. All three warnings fixed during review. Tests are comprehensive and follow project conventions. No regressions detected. Two informational suggestions remain (non-blocking).
