+++ generate.go
@@ -6,10 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"maps"
-	"os"
 	"slices"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/zendev-sh/goai/provider"
@@ -148,454 +145,6 @@ type StepResult struct {
 
 	// Sources contains citations/references from this step.
 	Sources []provider.Source
-}
-
-// TextStream is a streaming text generation response.
-//
-// Callers must consume the stream (via Stream, TextStream, or Result) or cancel
-// the context. Discarding a TextStream without consuming leaks goroutines.
-//
-// It provides three consumption modes (Stream, TextStream, Result).
-// Stream() and TextStream() are mutually exclusive - only call one.
-// Result() can always be called, including after Stream() or TextStream(),
-// to get the accumulated final result.
-type TextStream struct {
-	ctx           context.Context
-	source        <-chan provider.StreamChunk
-	consumeOnce   sync.Once
-	doneCh        chan struct{}
-	timeoutCancel context.CancelFunc
-
-	// Channels returned by the first Stream()/TextStream() call.
-	// Subsequent calls return the same channel instead of a dead one.
-	rawCh  <-chan provider.StreamChunk
-	textCh <-chan string
-
-	// Hook support.
-	onResponse   []func(ResponseInfo)
-	onStepFinish []func(StepResult)
-	onFinish     []func(FinishInfo)
-	onPanic      []func(PanicInfo)
-	startTime    time.Time
-
-	// stateRef, when non-nil, is transitioned to StepIdle when the consume
-	// goroutine returns. Only set by the single-shot StreamText path
-	// (streamWithToolLoop owns its own StateRef lifecycle inline). A nil
-	// stateRef is a no-op; AgentState.set handles nil receiver safely.
-	// See FIX 34.
-	stateRef *AgentState
-
-	// Accumulated state (written by consume goroutine, read after doneCh closes).
-	text             strings.Builder
-	toolCalls        []provider.ToolCall
-	sources          []provider.Source
-	finishReason     provider.FinishReason
-	usage            provider.Usage
-	response         provider.ResponseMetadata
-	providerMetadata map[string]map[string]any
-	streamErr        error
-
-	// Multi-step accumulation (written by consume goroutine).
-	steps         []StepResult
-	currentStep   int
-	stepText      strings.Builder
-	stepToolCalls []provider.ToolCall
-	stepSources   []provider.Source
-	reasoningBuf  strings.Builder // consolidated reasoning text (matches drainStep)
-	reasoningMeta map[string]any  // merged reasoning metadata (e.g. Anthropic signature)
-
-	// responseMessages is set by the streamWithToolLoop goroutine before doneCh closes.
-	responseMessages []provider.Message
-	// stepsExhausted is set by the streamWithToolLoop goroutine when MaxSteps was reached.
-	stepsExhausted bool
-}
-
-func newTextStream(ctx context.Context, source <-chan provider.StreamChunk) *TextStream {
-	return &TextStream{
-		ctx:    ctx,
-		source: source,
-		doneCh: make(chan struct{}),
-	}
-}
-
-// Stream returns a channel that emits raw StreamChunks from the provider.
-// Mutually exclusive with TextStream() - only call one streaming method.
-func (ts *TextStream) Stream() <-chan provider.StreamChunk {
-	ch := make(chan provider.StreamChunk, 64)
-	ts.consumeOnce.Do(func() {
-		ts.rawCh = ch
-		go ts.consume(ch, nil)
-	})
-	if ts.rawCh != nil {
-		return ts.rawCh
-	}
-	// Called after TextStream() consumed the source - return closed channel.
-	close(ch)
-	return ch
-}
-
-// TextStream returns the underlying channel of text chunks.
-// Note: this method has the same name as the containing type (TextStream);
-// call it as stream.TextStream() to receive the channel.
-// Mutually exclusive with Stream() - only call one streaming method.
-func (ts *TextStream) TextStream() <-chan string {
-	ch := make(chan string, 64)
-	ts.consumeOnce.Do(func() {
-		ts.textCh = ch
-		go ts.consume(nil, ch)
-	})
-	if ts.textCh != nil {
-		return ts.textCh
-	}
-	// Called after Stream() consumed the source - return closed channel.
-	close(ch)
-	return ch
-}
-
-// Result blocks until the stream completes and returns the accumulated result.
-// Check Err() after Result() to detect stream errors - Result does not surface
-// errors directly (use Err or check result.Steps for partial data).
-// Can be called after Stream() or TextStream() to get accumulated data.
-// Note: unlike ObjectStream.Result(), this method does not return an error.
-// Call Err() after Result() to check for stream errors.
-func (ts *TextStream) Result() *TextResult {
-	ts.consumeOnce.Do(func() {
-		go ts.consume(nil, nil)
-	})
-	<-ts.doneCh
-	return ts.buildResult()
-}
-
-// Err returns the first stream error encountered, or nil.
-// Must be called after the stream is fully consumed (after Result(),
-// or after the Stream()/TextStream() channel is drained).
-// Follows the bufio.Scanner.Err() pattern.
-func (ts *TextStream) Err() error {
-	<-ts.doneCh
-	return ts.streamErr
-}
-
-func (ts *TextStream) consume(rawOut chan<- provider.StreamChunk, textOut chan<- string) {
-	defer close(ts.doneCh)
-	// Surface a panic from any user hook fired during teardown (OnStepFinish,
-	// OnResponse, OnFinish) through stream.Err(). Registered near the top so it
-	// runs after (LIFO) the hook defers below, catching panics they re-raise via
-	// callHook. A pre-existing stream error is preserved as the root cause.
-	defer recoverToStreamErr(ts.onPanic, "stream", func(e error) {
-		if ts.streamErr == nil {
-			ts.streamErr = e
-		}
-	})
-	// FIX 34: transition StateRef to StepIdle as the consume goroutine
-	// exits. Single-shot StreamText wires this (stateRef is nil for the
-	// multi-step streamWithToolLoop path, which owns its own StateRef
-	// transitions inline). Deferred second from the top so it runs just
-	// before close(doneCh) - after all user hooks (OnResponse, OnStepFinish,
-	// OnFinish) have fired, so a poller that observes StepIdle can assume
-	// all hooks have already returned. Step count is 1 for a single-shot
-	// stream (the one DoStream call), regardless of whether the provider
-	// emitted any chunks.
-	if ts.stateRef != nil {
-		defer func() { ts.stateRef.set(StepIdle, 1) }()
-	}
-	if ts.timeoutCancel != nil {
-		defer ts.timeoutCancel()
-	}
-	if rawOut != nil {
-		defer close(rawOut)
-	}
-	if textOut != nil {
-		defer close(textOut)
-	}
-
-	// Call OnFinish hook when consume finishes (single-step streaming only).
-	// For multi-step streaming (streamWithToolLoop), OnFinish fires inline in the
-	// goroutine and ts.onFinish is nil, so this block is skipped.
-	// Deferred BEFORE OnStepFinish so it runs AFTER it (LIFO order).
-	if len(ts.onFinish) > 0 {
-		defer func() {
-			fireOnFinish(ts.onPanic, ts.onFinish, FinishInfo{
-				TotalSteps:   1,
-				TotalUsage:   ts.usage,
-				FinishReason: ts.finishReason,
-				StoppedBy:    provider.StopCauseNatural,
-			})
-		}()
-	}
-
-	// Call OnStepFinish hook when consume finishes (single-step streaming only).
-	// For multi-step streaming (streamWithToolLoop), OnStepFinish fires inline per
-	// step and ts.onStepFinish is nil, so this block is skipped.
-	// Deferred BEFORE OnResponse so it runs AFTER it (LIFO order), matching
-	// GenerateText's OnResponse → OnStepFinish sequence.
-	if len(ts.onStepFinish) > 0 {
-		defer func() {
-			stepResult := StepResult{
-				Number:           1,
-				Text:             ts.stepText.String(),
-				ToolCalls:        ts.toolCalls,
-				FinishReason:     ts.finishReason,
-				Usage:            ts.usage,
-				Response:         ts.response,
-				Sources:          ts.sources,
-				ProviderMetadata: ts.providerMetadata,
-			}
-			for _, fn := range ts.onStepFinish {
-				callHook(ts.onPanic, "OnStepFinish", func() { fn(stepResult) })
-			}
-		}()
-	}
-
-	// Call OnResponse hook when consume finishes (after all chunks processed).
-	if len(ts.onResponse) > 0 {
-		defer func() {
-			info := ResponseInfo{
-				Latency:      time.Since(ts.startTime),
-				Usage:        ts.usage,
-				FinishReason: ts.finishReason,
-				Error:        ts.streamErr,
-			}
-			var apiErr *APIError
-			if errors.As(ts.streamErr, &apiErr) {
-				info.StatusCode = apiErr.StatusCode
-			}
-			for _, fn := range ts.onResponse {
-				callHook(ts.onPanic, "OnResponse", func() { fn(info) })
-			}
-		}()
-	}
-
-	for chunk := range ts.source {
-		switch chunk.Type {
-		case provider.ChunkText:
-			ts.text.WriteString(chunk.Text)     // global accumulator (existing, includes reasoning)
-			ts.stepText.WriteString(chunk.Text) // per-step text-only accumulator (new)
-			if s, ok := chunk.Metadata["source"].(provider.Source); ok {
-				ts.sources = append(ts.sources, s)         // global (existing)
-				ts.stepSources = append(ts.stepSources, s) // per-step (new)
-			}
-
-		case provider.ChunkReasoning:
-			ts.text.WriteString(chunk.Text) // global accumulator (existing, includes reasoning)
-			// Consolidate reasoning fragments into one Part (matching drainStep behavior).
-			// Text is accumulated; metadata is merged (last chunk carries the signature).
-			if chunk.Text != "" {
-				ts.reasoningBuf.WriteString(chunk.Text)
-			}
-			if chunk.Metadata != nil {
-				if ts.reasoningMeta == nil {
-					ts.reasoningMeta = make(map[string]any)
-				}
-				for k, v := range chunk.Metadata {
-					ts.reasoningMeta[k] = v
-				}
-			}
-			if s, ok := chunk.Metadata["source"].(provider.Source); ok {
-				ts.sources = append(ts.sources, s)         // global (preserve existing behavior)
-				ts.stepSources = append(ts.stepSources, s) // per-step
-			}
-
-		case provider.ChunkToolCall:
-			tc := provider.ToolCall{ID: chunk.ToolCallID, Name: chunk.ToolName, Input: json.RawMessage(chunk.ToolInput), Metadata: chunk.Metadata}
-			ts.toolCalls = append(ts.toolCalls, tc)
-			ts.stepToolCalls = append(ts.stepToolCalls, tc)
-
-		case provider.ChunkStepFinish:
-			// GoAI-emitted tool-results backfill: update the last completed
-			// step's ToolResults (FIX 7 streaming parity). Emitted after
-			// executeToolsParallel returns so ts.steps exposes the same data
-			// the sync path's StepResult.ToolResults carries.
-			if stepSource, _ := chunk.Metadata["stepSource"].(string); stepSource == "goai-tool-results" {
-				if trs, ok := chunk.Metadata["toolResults"].([]provider.ToolResult); ok && len(ts.steps) > 0 {
-					ts.steps[len(ts.steps)-1].ToolResults = trs
-				}
-				continue
-			}
-			// GoAI-emitted step boundaries: build per-step StepResult.
-			if stepSource, _ := chunk.Metadata["stepSource"].(string); stepSource == "goai" {
-				ts.currentStep++
-				// Response is set directly on the chunk by the step loop.
-				// ProviderMetadata is embedded in Metadata (no dedicated StreamChunk field).
-				var stepProviderMeta map[string]map[string]any
-				if pm, ok := chunk.Metadata["providerMetadata"].(map[string]map[string]any); ok {
-					stepProviderMeta = pm
-				}
-				ts.steps = append(ts.steps, StepResult{
-					Number:           ts.currentStep,
-					Text:             ts.stepText.String(),
-					Reasoning:        ts.reasoningBuf.String(),
-					ToolCalls:        ts.stepToolCalls,
-					FinishReason:     chunk.FinishReason,
-					Usage:            chunk.Usage,
-					Sources:          ts.stepSources,
-					Response:         chunk.Response,
-					ProviderMetadata: stepProviderMeta,
-				})
-				// Accumulate per-step usage and finishReason for resilience:
-				// if stream terminates before ChunkFinish (e.g., context cancel between
-				// ChunkStepFinish and ChunkFinish sends), ts.usage still reflects completed steps.
-				// ChunkFinish will overwrite with the authoritative totalUsage when it arrives.
-				ts.usage = addUsage(ts.usage, chunk.Usage)
-				ts.finishReason = chunk.FinishReason
-				// Update ts.response for the overall TextResult (last step wins).
-				ts.response = chunk.Response
-				ts.providerMetadata = stepProviderMeta
-				// Reset per-step accumulators.
-				ts.stepText.Reset()
-				ts.stepToolCalls = nil
-				ts.stepSources = nil
-				ts.reasoningBuf.Reset()
-				ts.reasoningMeta = nil
-			} else {
-				// Provider-internal step boundary (e.g., Anthropic extended thinking).
-				// Preserve existing behavior: extract response, metadata, sources.
-				// This ensures single-step streaming continues to work correctly.
-				// NOTE: Do NOT accumulate usage here with addUsage. Provider-internal
-				// ChunkStepFinish usage is already included in the provider's ChunkFinish
-				// (which uses direct assignment), and the GoAI ChunkStepFinish (which uses
-				// addUsage via drainStep). Accumulating here would double-count.
-				ts.finishReason = chunk.FinishReason
-				ts.response = chunk.Response
-				if sources, ok := chunk.Metadata["sources"].([]provider.Source); ok {
-					ts.sources = append(ts.sources, sources...)
-					ts.stepSources = append(ts.stepSources, sources...)
-				}
-				if pm, ok := chunk.Metadata["providerMetadata"].(map[string]map[string]any); ok {
-					ts.providerMetadata = pm
-				}
-				// Copy flat metadata keys to Response.ProviderMetadata (existing behavior).
-				for k, v := range chunk.Metadata {
-					if k == "providerMetadata" || k == "sources" {
-						continue
-					}
-					if ts.response.ProviderMetadata == nil {
-						ts.response.ProviderMetadata = map[string]any{}
-					}
-					ts.response.ProviderMetadata[k] = v
-				}
-			}
-
-		case provider.ChunkFinish:
-			// Direct assignment (not addUsage): ChunkFinish carries authoritative total usage.
-			ts.usage = chunk.Usage
-			ts.finishReason = chunk.FinishReason
-			// Preserve existing single-step behavior: extract response, metadata, sources
-			// from the provider's ChunkFinish. For multi-step, the goai-emitted ChunkFinish
-			// does not carry these (they are embedded in ChunkStepFinish metadata instead).
-			if chunk.Response.ID != "" || chunk.Response.Model != "" {
-				ts.response = chunk.Response
-			}
-			if sources, ok := chunk.Metadata["sources"].([]provider.Source); ok {
-				ts.sources = append(ts.sources, sources...)
-			}
-			if pm, ok := chunk.Metadata["providerMetadata"].(map[string]map[string]any); ok {
-				ts.providerMetadata = pm
-			}
-			// Copy flat metadata keys to Response.ProviderMetadata (existing behavior).
-			for k, v := range chunk.Metadata {
-				if k == "providerMetadata" || k == "sources" {
-					continue
-				}
-				if ts.response.ProviderMetadata == nil {
-					ts.response.ProviderMetadata = map[string]any{}
-				}
-				ts.response.ProviderMetadata[k] = v
-			}
-
-		case provider.ChunkError:
-			if ts.streamErr == nil {
-				ts.streamErr = chunk.Error
-			}
-		}
-
-		if rawOut != nil {
-			select {
-			case rawOut <- chunk:
-			case <-ts.ctx.Done():
-				ts.streamErr = ts.ctx.Err()
-				return
-			}
-		}
-		if textOut != nil && chunk.Type == provider.ChunkText {
-			select {
-			case textOut <- chunk.Text:
-			case <-ts.ctx.Done():
-				ts.streamErr = ts.ctx.Err()
-				return
-			}
-		}
-		if rawOut == nil && textOut == nil {
-			if ts.ctx.Err() != nil {
-				ts.streamErr = ts.ctx.Err()
-				return
-			}
-		}
-	}
-}
-
-func (ts *TextStream) buildResult() *TextResult {
-	text := ts.text.String() // full accumulated text across all steps
-	result := &TextResult{
-		Text:             text,
-		ToolCalls:        ts.toolCalls,
-		FinishReason:     ts.finishReason,
-		TotalUsage:       ts.usage,
-		Response:         ts.response,
-		Sources:          ts.sources,
-		ProviderMetadata: ts.providerMetadata,
-	}
-	if len(ts.steps) > 0 {
-		result.Steps = ts.steps
-		// Match GenerateText: ToolCalls is the LAST step's tool calls, not all steps'.
-		result.ToolCalls = ts.steps[len(ts.steps)-1].ToolCalls
-		// Aggregate per-step reasoning so streaming TextResult exposes
-		// the same .Reasoning field as GenerateText.
-		var reasoningAll strings.Builder
-		for _, s := range ts.steps {
-			reasoningAll.WriteString(s.Reasoning)
-		}
-		result.Reasoning = reasoningAll.String()
-	} else if text != "" || len(ts.toolCalls) > 0 || ts.finishReason != "" {
-		// Single-step fallback (no multi-step ChunkStepFinish received, but data exists).
-		stepReasoning := ts.reasoningBuf.String()
-		result.Steps = []StepResult{{
-			Number:           1,
-			Text:             ts.stepText.String(),
-			Reasoning:        stepReasoning,
-			ToolCalls:        ts.toolCalls,
-			FinishReason:     ts.finishReason,
-			Usage:            ts.usage,
-			Response:         ts.response,
-			Sources:          ts.sources,
-			ProviderMetadata: ts.providerMetadata,
-		}}
-		result.Reasoning = stepReasoning
-	}
-	// No data: Steps is nil.
-
-	// Set StepsExhausted from streamWithToolLoop goroutine.
-	result.StepsExhausted = ts.stepsExhausted
-
-	// Populate ResponseMessages.
-	if ts.responseMessages != nil {
-		// Multi-step: set by streamWithToolLoop goroutine.
-		result.ResponseMessages = ts.responseMessages
-	} else if text != "" || len(result.ToolCalls) > 0 {
-		// Single-step: build a simple assistant message from the result.
-		// Use stepText (text-only, excludes reasoning) for ResponseMessages so reasoning
-		// doesn't get baked into PartText. Pass consolidated reasoning part separately
-		// (matching drainStep: one Part with merged metadata including signatures).
-		var reasoning []provider.Part
-		if ts.reasoningBuf.Len() > 0 || len(ts.reasoningMeta) > 0 {
-			reasoning = []provider.Part{{
-				Type:            provider.PartReasoning,
-				Text:            ts.reasoningBuf.String(),
-				ProviderOptions: ts.reasoningMeta,
-			}}
-		}
-		result.ResponseMessages = buildFinalAssistantMessages(ts.stepText.String(), result.ToolCalls, reasoning)
-	}
-	return result
 }
 
 // buildParams converts options to provider.GenerateParams.
@@ -764,8 +313,7 @@ func streamWithToolLoop(ctx context.Context, model provider.LanguageModel, o opt
 		// Both writes together ensure the invariant "highestInflightStep
 		// reflects the latest StepLLMInFlight announcement" holds on every
 		// path the defer can fire from.
-		highestInflightStep := 0
-		highestInflightStep = 1
+		highestInflightStep := 1
 		// Ensure any exit path (break, return, panic-recover-above, natural
 		// termination) leaves the observable state as Idle. Use closure-captured
 		// steps so the final step count is visible to pollers.
@@ -1185,7 +733,7 @@ func GenerateText(ctx context.Context, model provider.LanguageModel, opts ...Opt
 	// advertised StepLLMInFlight at N, so the final StepIdle must carry
 	// max(len(steps), highestInflightStep) to avoid a step-counter
 	// regression visible to pollers.
-	var highestInflightStep int
+	highestInflightStep := 1
 	// AgentState: initial (StepStarting, 0). set() is a no-op when o.StateRef is nil.
 	o.StateRef.set(StepStarting, 0)
 	defer func() {
@@ -1412,51 +960,6 @@ func GenerateText(ctx context.Context, model provider.LanguageModel, opts ...Opt
 	return tr, nil
 }
 
-// fireOnFinish calls all OnFinish hooks with individual panic recovery.
-// stopSafe evaluates a StopCondition with recover. A panicking predicate is
-// treated as "do not stop" and logged to stderr (consistent with how
-// OnBeforeStep / OnStepFinish handle panics).
-// finalizeStopCause classifies the terminal StopCause when the tool loop
-// exits by natural termination (no break set a cause). It encapsulates the
-// post-loop MaxSteps-exhaustion guard and the StopCauseNatural default so
-// both GenerateText and streamWithToolLoop share one implementation.
-//
-// Returns the resolved StopCause and whether MaxSteps was exhausted.
-func finalizeStopCause(hookStopped bool, current provider.StopCause, steps []StepResult, maxSteps int) (provider.StopCause, bool) {
-	stepsExhausted := false
-	if !hookStopped && current == "" && len(steps) >= maxSteps && len(steps) > 0 && len(steps[len(steps)-1].ToolCalls) > 0 {
-		stepsExhausted = true
-		current = provider.StopCauseMaxSteps
-	}
-	if current == "" {
-		current = provider.StopCauseNatural
-	}
-	return current, stepsExhausted
-}
-
-func stopSafe(onPanic []func(PanicInfo), pred StopCondition, steps []StepResult) bool {
-	var stop bool
-	// Pass a shallow defensive copy so predicates cannot re-order or truncate
-	// the internal slice. NOTE: this is a TOP-LEVEL copy only - nested slices
-	// (StepResult.ToolCalls, StepResult.ToolResults, StepResult.Content) are
-	// ALIASED into the caller's view. Predicates MUST treat the StepResult
-	// contents as read-only; mutating nested slices corrupts goai internal
-	// state and is a contract violation. Deep-clone would be prohibitively
-	// expensive per-step for a feature (predicate side-effects) that is not
-	// a supported use case.
-	//
-	// A panic in the predicate is fired to OnPanic and surfaced as a
-	// *PanicError (returned by GenerateText, or via stream.Err() for StreamText).
-	callHook(onPanic, "StopWhen", func() { stop = pred(slices.Clone(steps)) })
-	return stop
-}
-
-func fireOnFinish(onPanic []func(PanicInfo), hooks []func(FinishInfo), info FinishInfo) {
-	for _, fn := range hooks {
-		callHook(onPanic, "OnFinish", func() { fn(info) })
-	}
-}
-
 // requestMessages returns msgs with a system message prepended when system is non-empty.
 // Always allocates a new slice so hooks cannot mutate the caller's message state.
 func requestMessages(system string, msgs []provider.Message) []provider.Message {
@@ -1468,714 +971,6 @@ func requestMessages(system string, msgs []provider.Message) []provider.Message 
 	out := make([]provider.Message, 0, len(msgs)+1)
 	out = append(out, SystemMessage(system))
 	return append(out, msgs...)
-}
-
-// isProviderExecuted reports whether a tool call was executed by the provider
-// (e.g. Anthropic web_search, OpenAI web_search_call). The provider returns
-// the result inline on the assistant turn, so goai must not look it up in
-// the user's tool map or synthesize a tool-result message for it.
-//
-// Providers mark these calls by setting tc.Metadata["providerExecuted"] = true
-// (or by attaching a non-nil "resultBlock" / "rawItem" payload that the
-// request serializer re-emits verbatim).
-func isProviderExecuted(tc provider.ToolCall) bool {
-	if tc.Metadata == nil {
-		return false
-	}
-	if v, ok := tc.Metadata["providerExecuted"].(bool); ok && v {
-		return true
-	}
-	if _, ok := tc.Metadata["resultBlock"]; ok {
-		return true
-	}
-	if _, ok := tc.Metadata["rawItem"]; ok {
-		return true
-	}
-	return false
-}
-
-// buildToolMap creates a name→Tool lookup of executable tools from the options.
-//
-// Tool names must be unique across all tools, not just executable ones: the
-// Vercel AI SDK (our reference) keys its tool set by name, so a collision there
-// is impossible. We validate every named tool to match that guarantee, while
-// the returned map only contains tools that have an Execute function.
-func buildToolMap(tools []Tool) (map[string]Tool, error) {
-	if len(tools) == 0 {
-		return nil, nil
-	}
-	m := make(map[string]Tool, len(tools))
-	seen := make(map[string]struct{}, len(tools))
-	for _, t := range tools {
-		if t.Name == "" {
-			if t.Execute != nil {
-				fmt.Fprintf(os.Stderr, "goai: tool with empty name skipped\n")
-			}
-			continue
-		}
-		if _, exists := seen[t.Name]; exists {
-			return nil, fmt.Errorf("goai: duplicate tool name %q: tool names must be unique", t.Name)
-		}
-		seen[t.Name] = struct{}{}
-		if t.Execute != nil {
-			m[t.Name] = t
-		}
-	}
-	if len(m) == 0 {
-		return nil, nil
-	}
-	return m, nil
-}
-
-// toolOutput holds the result of a single tool execution (package-level type
-// shared between executeToolsParallel and buildToolMessages).
-type toolOutput struct {
-	index  int
-	result string
-	err    error
-}
-
-// toolHooks bundles the hook functions and options passed to executeToolsParallel.
-type toolHooks struct {
-	sequential      bool // when true, execute tools one at a time
-	onToolCallStart []func(ToolCallStartInfo)
-	onToolCall      []func(ToolCallInfo)
-	onBeforeExecute func(BeforeToolExecuteInfo) BeforeToolExecuteResult
-	onAfterExecute  func(AfterToolExecuteInfo) AfterToolExecuteResult
-	onPanic         []func(PanicInfo)
-}
-
-func executeToolsParallel(
-	ctx context.Context,
-	calls []provider.ToolCall,
-	toolMap map[string]Tool,
-	step int,
-	hooks toolHooks,
-) ([]provider.Message, []provider.ToolResult) {
-
-	results := make([]toolOutput, len(calls))
-	var wg sync.WaitGroup
-
-	for i, tc := range calls {
-		// Server-executed tool calls (e.g. Anthropic web_search, OpenAI
-		// web_search_call) are run by the provider itself; their results are
-		// already inline on the assistant turn. They have no Execute and must
-		// not be looked up in toolMap or surfaced as unknown-tool errors.
-		if isProviderExecuted(tc) {
-			results[i] = toolOutput{index: i, result: ""}
-			continue
-		}
-		tool, ok := toolMap[tc.Name]
-		if !ok {
-			// Unknown tool: fire OnToolCallStart + OnToolCall with ErrUnknownTool.
-			// Each hook is independently recover-wrapped so a panic in OnToolCallStart
-			// does not prevent OnToolCall from firing.
-			//
-			// Asymmetry with known-tool path (below): for known tools, OnToolCallStart
-			// panic prevents Execute from running (the tool should not execute if the
-			// pre-hook crashed). For unknown tools, Execute never runs anyway, so both
-			// hooks fire independently for observability completeness.
-			results[i] = toolOutput{index: i, err: ErrUnknownTool}
-			for _, fn := range hooks.onToolCallStart {
-				func(f func(ToolCallStartInfo)) {
-					defer func() {
-						if r := recover(); r != nil {
-							firePanic(hooks.onPanic, "OnToolCallStart", r)
-						}
-					}()
-					f(ToolCallStartInfo{ToolCallID: tc.ID, ToolName: tc.Name, Step: step, Input: tc.Input})
-				}(fn)
-			}
-			for _, fn := range hooks.onToolCall {
-				func(f func(ToolCallInfo)) {
-					defer func() {
-						if r := recover(); r != nil {
-							firePanic(hooks.onPanic, "OnToolCall", r)
-						}
-					}()
-					f(ToolCallInfo{
-						ToolCallID: tc.ID,
-						ToolName:   tc.Name,
-						Step:       step,
-						Input:      tc.Input,
-						Error:      ErrUnknownTool,
-					})
-				}(fn)
-			}
-			continue
-		}
-
-		toolFn := func(i int, tc provider.ToolCall, tool Tool) {
-			if !hooks.sequential {
-				defer wg.Done()
-			}
-			var hookFired bool // true after OnToolCallStart completes (before Execute)
-			var executed bool  // tracks whether Execute ran (for panic recovery)
-			defer func() {
-				if r := recover(); r != nil {
-					if !executed {
-						// The full panic value + stack go to OnPanic (firePanic);
-						// the tool error sent to the LLM omits the raw value so a
-						// panic carrying sensitive data is not echoed to the model.
-						if !hookFired {
-							firePanic(hooks.onPanic, "OnToolCallStart", r)
-							results[i] = toolOutput{index: i, err: fmt.Errorf("goai: OnToolCallStart hook for tool %q panicked", tc.Name)}
-						} else {
-							firePanic(hooks.onPanic, "tool:"+tc.Name, r)
-							results[i] = toolOutput{index: i, err: fmt.Errorf("goai: tool %q panicked", tc.Name)}
-						}
-					}
-					// executed==true: Execute succeeded, results[i] already set.
-					// OnToolCall panic after Execute is swallowed (preserve result).
-				}
-			}()
-
-			// OnToolCallStart: pre-execution (each independently recover-wrapped).
-			var hookPanicked bool
-			for _, fn := range hooks.onToolCallStart {
-				func(f func(ToolCallStartInfo)) {
-					defer func() {
-						if r := recover(); r != nil {
-							firePanic(hooks.onPanic, "OnToolCallStart", r)
-							hookPanicked = true
-						}
-					}()
-					f(ToolCallStartInfo{
-						ToolCallID: tc.ID,
-						ToolName:   tc.Name,
-						Step:       step,
-						Input:      tc.Input,
-					})
-				}(fn)
-			}
-			hookFired = true
-			if hookPanicked {
-				panicStr := fmt.Sprintf("goai: OnToolCallStart hook for tool %q panicked", tc.Name)
-				results[i] = toolOutput{index: i, err: fmt.Errorf("%s", panicStr)}
-				return
-			}
-
-			// Create tool context early so both hooks and Execute share it.
-			toolCtx := context.WithValue(ctx, toolCallIDKey{}, tc.ID)
-
-			// OnBeforeToolExecute: can skip execution (permission, doom loop, etc.).
-			if hooks.onBeforeExecute != nil {
-				var beforeResult BeforeToolExecuteResult
-				func() {
-					defer func() {
-						if r := recover(); r != nil {
-							firePanic(hooks.onPanic, "OnBeforeToolExecute", r)
-							// Raw value goes to OnPanic only; omit it here so a
-							// panic with sensitive data is not echoed to the model.
-							beforeResult = BeforeToolExecuteResult{
-								Skip:  true,
-								Error: errors.New("goai: OnBeforeToolExecute hook panicked"),
-							}
-						}
-					}()
-					beforeResult = hooks.onBeforeExecute(BeforeToolExecuteInfo{
-						Ctx:        toolCtx,
-						ToolCallID: tc.ID,
-						ToolName:   tc.Name,
-						Step:       step,
-						Input:      tc.Input,
-					})
-				}()
-				if beforeResult.Skip {
-					executed = true // prevent outer panic handler from overwriting
-					if beforeResult.Error != nil {
-						results[i] = toolOutput{index: i, result: beforeResult.Result, err: beforeResult.Error}
-					} else {
-						results[i] = toolOutput{index: i, result: beforeResult.Result}
-					}
-					// Fire OnToolCall with skipped result for observability.
-					for _, fn := range hooks.onToolCall {
-						func(f func(ToolCallInfo)) {
-							defer func() {
-								if r := recover(); r != nil {
-									firePanic(hooks.onPanic, "OnToolCall", r)
-								}
-							}()
-							f(ToolCallInfo{
-								ToolCallID: tc.ID,
-								ToolName:   tc.Name,
-								Step:       step,
-								Input:      tc.Input,
-								Output:     beforeResult.Result,
-								StartTime:  time.Now(),
-								Skipped:    true,
-								Error:      beforeResult.Error,
-							})
-						}(fn)
-					}
-					return
-				}
-				// Apply hook overrides for non-skipped tools.
-				if beforeResult.Ctx != nil {
-					toolCtx = beforeResult.Ctx
-				}
-				if beforeResult.Input != nil {
-					tc.Input = beforeResult.Input
-				}
-			}
-
-			start := time.Now()
-			output, err := tool.Execute(toolCtx, tc.Input)
-			executed = true
-
-			// OnAfterToolExecute: can modify output (secret scanning, truncation, etc.).
-			var afterMetadata map[string]any
-			if hooks.onAfterExecute != nil {
-				func() {
-					defer func() {
-						if r := recover(); r != nil {
-							firePanic(hooks.onPanic, "OnAfterToolExecute", r)
-							// Preserve original result on panic.
-						}
-					}()
-					afterResult := hooks.onAfterExecute(AfterToolExecuteInfo{
-						Ctx:        toolCtx,
-						ToolCallID: tc.ID,
-						ToolName:   tc.Name,
-						Step:       step,
-						Input:      tc.Input,
-						Output:     output,
-						Error:      err,
-					})
-					if afterResult.Output != "" {
-						output = afterResult.Output
-					}
-					if afterResult.Error != nil {
-						err = afterResult.Error
-					}
-					afterMetadata = afterResult.Metadata
-				}()
-			}
-
-			results[i] = toolOutput{index: i, result: output, err: err}
-
-			// OnToolCall: post-execution (each independently recover-wrapped).
-			for _, fn := range hooks.onToolCall {
-				func(f func(ToolCallInfo)) {
-					defer func() {
-						if r := recover(); r != nil {
-							firePanic(hooks.onPanic, "OnToolCall", r)
-						}
-					}()
-					info := ToolCallInfo{
-						ToolCallID: tc.ID,
-						ToolName:   tc.Name,
-						Step:       step,
-						Input:      tc.Input,
-						Output:     output,
-						StartTime:  start,
-						Duration:   time.Since(start),
-						Error:      err,
-						Metadata:   afterMetadata,
-					}
-					var parsed any
-					if err == nil && json.Unmarshal([]byte(output), &parsed) == nil {
-						info.OutputObject = parsed
-					}
-					f(info)
-				}(fn)
-			}
-		}
-		if hooks.sequential {
-			toolFn(i, tc, tool)
-		} else {
-			wg.Add(1)
-			go toolFn(i, tc, tool)
-		}
-	}
-
-	if !hooks.sequential {
-		wg.Wait()
-	}
-	return buildToolMessages(calls, results), buildToolResults(calls, results)
-}
-
-// buildToolResults converts raw tool outputs into structured provider.ToolResult
-// values (one per call, in call order). The Output string mirrors exactly what
-// buildToolMessages places in the tool-result message content so consumers can
-// correlate predicate input with the on-wire transcript.
-func buildToolResults(calls []provider.ToolCall, results []toolOutput) []provider.ToolResult {
-	if len(calls) == 0 {
-		return nil
-	}
-	out := make([]provider.ToolResult, len(calls))
-	for i, tc := range calls {
-		r := results[i]
-		tr := provider.ToolResult{
-			ToolCallID: tc.ID,
-			ToolName:   tc.Name,
-			Error:      r.err,
-			IsError:    r.err != nil,
-		}
-		if r.err != nil {
-			errStr := r.err.Error()
-			runes := []rune(errStr)
-			if len(runes) > 500 {
-				errStr = string(runes[:500]) + "..."
-			}
-			tr.Output = "error: " + errStr
-		} else {
-			tr.Output = r.result
-		}
-		out[i] = tr
-	}
-	return out
-}
-
-// buildToolMessages converts tool call results to provider messages.
-// Note: toolOutput is defined as a package-level type (not function-scoped)
-// so both executeToolsParallel and buildToolMessages can reference it.
-func buildToolMessages(calls []provider.ToolCall, results []toolOutput) []provider.Message {
-	msgs := make([]provider.Message, 0, len(calls))
-	for i, tc := range calls {
-		// Server-executed calls have no separate tool message; their result
-		// is delivered inline on the assistant turn.
-		if isProviderExecuted(tc) {
-			continue
-		}
-		r := results[i]
-		if r.err != nil {
-			errStr := r.err.Error()
-			runes := []rune(errStr)
-			if len(runes) > 500 {
-				errStr = string(runes[:500]) + "..."
-			}
-			msgs = append(msgs, ToolMessage(tc.ID, tc.Name, "error: "+errStr))
-		} else {
-			msgs = append(msgs, ToolMessage(tc.ID, tc.Name, r.result))
-		}
-	}
-	return msgs
-}
-
-// appendToolRoundTrip appends an assistant message (with tool_use parts)
-// and tool result messages for the streaming tool loop.
-// reasoning parts are placed first so providers that require thinking blocks
-// (e.g. Bedrock with extended thinking) see them before tool_use content.
-func appendToolRoundTrip(
-	msgs []provider.Message,
-	text string,
-	reasoning []provider.Part,
-	toolCalls []provider.ToolCall,
-	toolMsgs []provider.Message,
-) []provider.Message {
-	var parts []provider.Part
-	// Reasoning first (before text and tool_use). Clone ProviderOptions to avoid
-	// aliasing between params.Messages and ResponseMessages.
-	for _, r := range reasoning {
-		parts = append(parts, provider.Part{
-			Type:            r.Type,
-			Text:            r.Text,
-			ProviderOptions: maps.Clone(r.ProviderOptions),
-		})
-	}
-	if text != "" {
-		parts = append(parts, provider.Part{Type: provider.PartText, Text: text})
-	}
-	for _, tc := range toolCalls {
-		parts = append(parts, provider.Part{
-			Type:            provider.PartToolCall,
-			ToolCallID:      tc.ID,
-			ToolName:        tc.Name,
-			ToolInput:       append(json.RawMessage(nil), tc.Input...), // clone byte slice
-			ProviderOptions: maps.Clone(tc.Metadata),                   // shallow clone (matches buildFinalAssistantMessages)
-		})
-	}
-	msgs = append(msgs, provider.Message{Role: provider.RoleAssistant, Content: parts})
-	msgs = append(msgs, toolMsgs...)
-	return msgs
-}
-
-// buildTextResult constructs a TextResult from accumulated steps.
-// Caller must ensure steps is non-empty.
-func buildTextResult(steps []StepResult, totalUsage provider.Usage) *TextResult {
-	if len(steps) == 0 {
-		return &TextResult{TotalUsage: totalUsage}
-	}
-	last := steps[len(steps)-1]
-	// Accumulate text from all steps.
-	var allText strings.Builder
-	for _, s := range steps {
-		allText.WriteString(s.Text)
-	}
-	// Accumulate reasoning text from all steps. Concatenated as-is so
-	// callers see the same boundaries the steps had; consumers wanting
-	// per-step reasoning can iterate Steps directly.
-	var allReasoning strings.Builder
-	for _, s := range steps {
-		allReasoning.WriteString(s.Reasoning)
-	}
-	// Collect sources from all steps.
-	var allSources []provider.Source
-	for _, s := range steps {
-		allSources = append(allSources, s.Sources...)
-	}
-
-	return &TextResult{
-		Text:             allText.String(),
-		Reasoning:        allReasoning.String(),
-		ToolCalls:        last.ToolCalls,
-		Steps:            steps,
-		TotalUsage:       totalUsage,
-		FinishReason:     last.FinishReason,
-		Response:         last.Response,
-		ProviderMetadata: last.ProviderMetadata,
-		Sources:          allSources,
-	}
-}
-
-// buildResponseMessages constructs the full ResponseMessages from the tool round-trip
-// messages (delta between original and final params.Messages) and the steps.
-//
-// With Vercel-parity StopWhen placement (evaluated AFTER tool execution and
-// appendToolRoundTrip), the delta always contains the assistant + tool-result
-// messages for every completed step whose LLM response carried tool calls --
-// including the last step on a StopWhen break and at MaxSteps exhaustion. The
-// only case the delta is missing the last step's assistant message is a
-// natural termination (last step produced text with no tool calls): that
-// message is appended here.
-//
-// Consecutive tool messages are merged into a single message because some
-// providers require parallel tool results in a single message (not split
-// across multiple messages).
-//
-// The reasoning parameter provides thinking/reasoning parts for the final
-// assistant message (streaming path only; GenerateText passes nil). It is
-// only applied when the delta is empty or the last step had no tool calls
-// (i.e. when we are actually building the final assistant message here).
-func buildResponseMessages(roundTripDelta []provider.Message, steps []StepResult, reasoning []provider.Part) []provider.Message {
-	if len(steps) == 0 {
-		return nil
-	}
-	last := steps[len(steps)-1]
-	if len(roundTripDelta) == 0 {
-		return buildFinalAssistantMessages(last.Text, last.ToolCalls, reasoning)
-	}
-	msgs := mergeToolMessages(roundTripDelta)
-	if len(last.ToolCalls) > 0 {
-		// Delta already contains this step's assistant + tool-result messages
-		// (appendToolRoundTrip ran before loop break). Avoid duplication.
-		return msgs
-	}
-	// Natural termination: last step produced text with no tool calls - its
-	// assistant message is NOT yet in the delta.
-	finalMsg := buildFinalAssistantMessages(last.Text, last.ToolCalls, reasoning)
-	return append(msgs, finalMsg...)
-}
-
-// mergeToolMessages merges consecutive tool-role messages into single messages.
-// The internal tool loop creates one message per tool call, but callers expect
-// parallel tool results grouped in a single message per round-trip.
-func mergeToolMessages(msgs []provider.Message) []provider.Message {
-	out := make([]provider.Message, 0, len(msgs))
-	for _, m := range msgs {
-		if m.Role == provider.RoleTool && len(out) > 0 && out[len(out)-1].Role == provider.RoleTool {
-			// Merge parts into the previous tool message.
-			out[len(out)-1].Content = append(out[len(out)-1].Content, m.Content...)
-		} else {
-			// Clone the message to avoid mutating the original.
-			out = append(out, provider.Message{
-				Role:    m.Role,
-				Content: slices.Clone(m.Content),
-			})
-		}
-	}
-	return out
-}
-
-// buildFinalAssistantMessages builds a single assistant message from text, tool calls,
-// and/or reasoning parts. Returns nil when all inputs are empty.
-// Reasoning parts are placed first so providers that require thinking blocks
-// (e.g. Bedrock with extended thinking) see them before text/tool_use content.
-func buildFinalAssistantMessages(text string, toolCalls []provider.ToolCall, reasoning []provider.Part) []provider.Message {
-	var parts []provider.Part
-	parts = append(parts, reasoning...)
-	if text != "" {
-		parts = append(parts, provider.Part{Type: provider.PartText, Text: text})
-	}
-	for _, tc := range toolCalls {
-		parts = append(parts, provider.Part{
-			Type:       provider.PartToolCall,
-			ToolCallID: tc.ID,
-			ToolName:   tc.Name,
-			ToolInput:  append(json.RawMessage(nil), tc.Input...),
-			// Shallow copy of Metadata , matches the existing appendToolRoundTrip pattern.
-			// Nested map values are shared, not deep-cloned.
-			ProviderOptions: maps.Clone(tc.Metadata),
-		})
-	}
-	if len(parts) == 0 {
-		return nil
-	}
-	return []provider.Message{{Role: provider.RoleAssistant, Content: parts}}
-}
-
-type drainResult struct {
-	text             string          // text-only (ChunkText), used for appendToolRoundTrip
-	reasoning        []provider.Part // reasoning/thinking parts, echoed back for providers that require it (e.g. Bedrock)
-	reasoningText    string          // consolidated reasoning text (PartReasoning), surfaced via StepResult.Reasoning
-	toolCalls        []provider.ToolCall
-	usage            provider.Usage
-	finishReason     provider.FinishReason
-	sources          []provider.Source
-	response         provider.ResponseMetadata
-	providerMetadata map[string]map[string]any
-	err              error // non-nil if context cancelled during drain
-}
-
-func drainStep(
-	ctx context.Context,
-	source <-chan provider.StreamChunk,
-	out chan<- provider.StreamChunk,
-) drainResult {
-	var (
-		textBuf       strings.Builder // ChunkText only (reasoning excluded)
-		reasoningBuf  strings.Builder // accumulated reasoning text
-		reasoningMeta map[string]any  // last metadata (contains signature)
-		dr            drainResult
-	)
-
-	for chunk := range source {
-		// Forward chunk to consumer. Suppress these types (handled explicitly below):
-		// - ChunkFinish: step loop emits its own ChunkFinish with totalUsage
-		// - ChunkError: forwarded explicitly in the switch to avoid double-send
-		// Provider ChunkStepFinish IS forwarded (not suppressed) so Stream() consumers
-		// can see provider-internal boundaries (e.g., Anthropic thinking steps). These
-		// do NOT carry Metadata["stepSource"]="goai", so consume() distinguishes them.
-		if chunk.Type != provider.ChunkFinish && chunk.Type != provider.ChunkError {
-			if !provider.TrySend(ctx, out, chunk) {
-				// Drain source to unblock provider.
-				drainRemaining(source)
-				dr.err = ctx.Err()
-				return dr
-			}
-		}
-
-		// Accumulate state (same logic as TextStream.consume, generate.go:198-242).
-		// Note: ChunkToolCallDelta, ChunkToolCallStreamStart, and ChunkToolResult are
-		// forwarded to the consumer (not suppressed) but NOT accumulated here. drainStep
-		// only captures complete ChunkToolCall chunks. Providers always emit a final
-		// ChunkToolCall with complete data after all deltas.
-		switch chunk.Type {
-		case provider.ChunkText:
-			textBuf.WriteString(chunk.Text)
-			if s, ok := chunk.Metadata["source"].(provider.Source); ok {
-				dr.sources = append(dr.sources, s)
-			}
-		case provider.ChunkReasoning:
-			// Accumulate into a single buffer. The final chunk carries the
-			// signature (text="", metadata={"signature":"..."}); earlier chunks
-			// carry text fragments. Consolidating produces one complete part.
-			if chunk.Text != "" {
-				reasoningBuf.WriteString(chunk.Text)
-			}
-			if chunk.Metadata != nil {
-				if reasoningMeta == nil {
-					reasoningMeta = make(map[string]any)
-				}
-				for k, v := range chunk.Metadata {
-					reasoningMeta[k] = v
-				}
-			}
-			if s, ok := chunk.Metadata["source"].(provider.Source); ok {
-				dr.sources = append(dr.sources, s)
-			}
-		case provider.ChunkToolCall:
-			dr.toolCalls = append(dr.toolCalls, provider.ToolCall{
-				ID:       chunk.ToolCallID,
-				Name:     chunk.ToolName,
-				Input:    json.RawMessage(chunk.ToolInput),
-				Metadata: chunk.Metadata,
-			})
-		case provider.ChunkStepFinish:
-			// Provider-internal step boundary (e.g., Anthropic extended thinking).
-			// Use direct assignment (last value wins), matching ChunkFinish below.
-			// This is correct for both zero-usage providers (Anthropic: 0 overwrites 0)
-			// and running-total providers (Google: last total is authoritative).
-			dr.finishReason = chunk.FinishReason
-			dr.usage = chunk.Usage
-			dr.response = chunk.Response
-			if sources, ok := chunk.Metadata["sources"].([]provider.Source); ok {
-				dr.sources = append(dr.sources, sources...)
-			}
-			if pm, ok := chunk.Metadata["providerMetadata"].(map[string]map[string]any); ok {
-				dr.providerMetadata = pm
-			}
-			for k, v := range chunk.Metadata {
-				if k == "providerMetadata" || k == "sources" {
-					continue
-				}
-				if dr.response.ProviderMetadata == nil {
-					dr.response.ProviderMetadata = map[string]any{}
-				}
-				dr.response.ProviderMetadata[k] = v
-			}
-		case provider.ChunkFinish:
-			// Terminal chunk. Use direct assignment for usage (not addUsage) to avoid
-			// double-counting when providers emit both ChunkStepFinish and ChunkFinish
-			// with the same accumulated usage (e.g., Google).
-			dr.finishReason = chunk.FinishReason
-			dr.usage = chunk.Usage
-			dr.response = chunk.Response
-			if sources, ok := chunk.Metadata["sources"].([]provider.Source); ok {
-				dr.sources = append(dr.sources, sources...)
-			}
-			if pm, ok := chunk.Metadata["providerMetadata"].(map[string]map[string]any); ok {
-				dr.providerMetadata = pm
-			}
-			// Copy flat metadata keys to Response.ProviderMetadata (same as consume(),
-			// generate.go:229-237). Providers use this for per-response data: Anthropic
-			// ("iterations", "contextManagement"), Bedrock ("cacheWriteInputTokens").
-			for k, v := range chunk.Metadata {
-				if k == "providerMetadata" || k == "sources" {
-					continue
-				}
-				if dr.response.ProviderMetadata == nil {
-					dr.response.ProviderMetadata = map[string]any{}
-				}
-				dr.response.ProviderMetadata[k] = v
-			}
-		case provider.ChunkError:
-			// Forward error chunks to consumer. Mid-stream errors flow through
-			// ChunkError chunks to the consumer; OnResponse does not report them.
-			if !provider.TrySend(ctx, out, chunk) {
-				drainRemaining(source)
-				dr.err = ctx.Err()
-				return dr
-			}
-		}
-	}
-
-	dr.text = textBuf.String()
-	dr.reasoningText = reasoningBuf.String()
-
-	// Consolidate reasoning fragments into one Part (text + signature metadata)
-	// so the codec serializes a single complete block.
-	if reasoningBuf.Len() > 0 || len(reasoningMeta) > 0 {
-		dr.reasoning = []provider.Part{{
-			Type:            provider.PartReasoning,
-			Text:            reasoningBuf.String(),
-			ProviderOptions: reasoningMeta,
-		}}
-	}
-
-	return dr
-}
-
-// drainRemaining reads and discards all remaining chunks from source.
-// This unblocks the provider's write-side goroutine on context cancellation,
-// preventing goroutine leaks. Note: this blocks until the provider closes its
-// channel. A misbehaving provider that never closes will cause this to hang.
-// All current providers use defer close(out) so this is safe in practice.
-func drainRemaining(source <-chan provider.StreamChunk) {
-	for range source {
-		// discard
-	}
 }
 
 // addUsage adds b's counts to a and returns the result.
